+++ app/docs/layout.tsx
@@ -1,16 +1,23 @@
 "use client";
 import { metadata } from "./metadata";
 
+import DocsSidebar from "@/components/docs-sidebar";
+
 export default function DocsLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
-    <section className="flex flex-col items-center justify-center gap-4 py-8 md:py-10">
-      <div className="inline-block max-w-lg justify-center text-center">
-        {children}
-      </div>
-    </section>
+    <div className="flex">
+      <DocsSidebar />
+      <main className="flex-1 p-4">
+        <section className="flex flex-col items-center justify-center gap-4 py-8 md:py-10">
+          <div className="inline-block max-w-lg justify-center text-center">
+            {children}
+          </div>
+        </section>
+      </main>
+    </div>
   );
 }