+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updated the requirements page with content from the README.md file.
+- Updated the about page with hardware specifications from the README.md file.
 
 ## [0.2.1] - 2024-04-28